Work around for getting linked object properties
Realm has what appears to be a bug in ListView calling renderRow with
some inputs it shouldn’t be at the start. Got a function for doing
item.department.name as it appears to only affect that.

Also editableCell shows on the table now, but editing not persisted at
all.