DropDown should work but doesn't!
RAAAGE.

Awkward moment when I don’t know if it doesn’t work because I have used
it slightly wrong or if a new version of react broke it sometime since
the last commit 5 months ago.

Doesn’t appear to call getCategoryOptionList() when the component is
clicked.

Preferably the above function and setCategory would just be arrow
functions within the JSX, but making them outside conforms closer to
the suggested implementation.